Add function to get ordinal suffix for dates
This commit adds a new function that can be used to get the ordinal suffix for a given day of the month. This function takes a number as input and returns a string with the appropriate ordinal suffix ("st", "nd", "rd", or "th"). This function is used in the `MyTime` component to display the date with the correct suffix.